// Chat-App/src/app/ChatAppModule/models/message.ts
export  class  Message {
    IdChat: number;
    UserName: string;
    Content_mess: string;
    PathFile: string;
    IsDelAll: boolean;
    IsHidenAll: boolean;
    IsVideoFile: boolean;
    isFile: boolean;
    isCall: boolean;
    TenFile: string;
    IdGroup: number;
    Note:string;
    Attachment:any[];
    isInsertMember:boolean;
    //dateRead?: Date;
    CreatedDate: Date;
}